requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
pydantic
httpx
passlib[bcrypt]
python-jose
